+++ shared/log.rs
@@ -6,14 +6,14 @@ use ::std::path::{Path, PathBuf};
 
 use crate::settings::Settings;
 
-pub fn get_log_paths(task_id: i32, settings: &Settings) -> (PathBuf, PathBuf) {
+pub fn get_log_paths(task_id: usize, settings: &Settings) -> (PathBuf, PathBuf) {
     let pueue_dir = Path::new(&settings.daemon.pueue_directory).join("temp");
     let out_path = pueue_dir.join(format!("{}_stdout.log", task_id));
     let err_path = pueue_dir.join(format!("{}_stderr.log", task_id));
     (out_path, err_path)
 }
 
-pub fn create_log_file_handles(task_id: i32, settings: &Settings) -> Result<(File, File)> {
+pub fn create_log_file_handles(task_id: usize, settings: &Settings) -> Result<(File, File)> {
     let (out_path, err_path) = get_log_paths(task_id, settings);
     let stdout = File::create(out_path)?;
     let stderr = File::create(err_path)?;
@@ -21,7 +21,7 @@ pub fn create_log_file_handles(task_id: i32, settings: &Settings) -> Result<(Fil
     Ok((stdout, stderr))
 }
 
-pub fn get_log_file_handles(task_id: i32, settings: &Settings) -> Result<(File, File)> {
+pub fn get_log_file_handles(task_id: usize, settings: &Settings) -> Result<(File, File)> {
     let (out_path, err_path) = get_log_paths(task_id, settings);
     let stdout = File::open(out_path)?;
     let stderr = File::open(err_path)?;
@@ -29,7 +29,7 @@ pub fn get_log_file_handles(task_id: i32, settings: &Settings) -> Result<(File, 
     Ok((stdout, stderr))
 }
 
-pub fn read_log_files(task_id: i32, settings: &Settings) -> Result<(String, String)> {
+pub fn read_log_files(task_id: usize, settings: &Settings) -> Result<(String, String)> {
     let (mut stdout_handle, mut stderr_handle) = get_log_file_handles(task_id, settings)?;
     let mut stdout = String::new();
     let mut stderr = String::new();
@@ -40,7 +40,7 @@ pub fn read_log_files(task_id: i32, settings: &Settings) -> Result<(String, Stri
     Ok((stdout, stderr))
 }
 
-pub fn clean_log_handles(task_id: i32, settings: &Settings) {
+pub fn clean_log_handles(task_id: usize, settings: &Settings) {
     let (out_path, err_path) = get_log_paths(task_id, settings);
     if let Err(err) = remove_file(out_path) {
         error!(
